move general, example plugins to pluginimpl directory
