+++ dce/main.go
@@ -30,8 +30,8 @@ import (
 	"github.com/paypal/dce-go/config"
 	"github.com/paypal/dce-go/dce/monitor"
 	"github.com/paypal/dce-go/plugin"
-	_ "github.com/paypal/dce-go/plugin/example"
-	_ "github.com/paypal/dce-go/plugin/general"
+	_ "github.com/paypal/dce-go/pluginimpl/example"
+	_ "github.com/paypal/dce-go/pluginimpl/general"
 
 	"github.com/paypal/dce-go/types"
 	"github.com/paypal/dce-go/utils"
