+++ docs/how-to-develop.md
@@ -81,7 +81,7 @@ PreLaunchTask and PostLaunchTask have Context object as first parameter. This is
 Below sample code illustrates a sample plugin loading parsed compose file from context in an object of ServiceDetail type.
 
 ##### Sample Plugin implementation
-Example plugin implementation can be found [here](../plugin/example/impl.go)
+Example plugin implementation can be found [here](../pluginimpl/example/impl.go)
 ```
 func (ex *exampleExt) PreLaunchTask(ctx *context.Context, composeFiles *[]string, executorId string, taskInfo *mesos.TaskInfo) error {
 	logger.Println("PreLaunchTask Starting")
@@ -156,7 +156,7 @@ func (ex *exampleExt) Shutdown(executor.ExecutorDriver) error {
    ```
       Here is an example.
    ```
-       _ "github.com/paypal/dce-go/plugin/example"
+       _ "github.com/paypal/dce-go/pluginimpl/example"
    ```
 3. Last but not least, an order of plugins need to be specified in the main configuration file([config.yaml](../config/config.yaml)). Plugins are automatically invoked in that order at appropriate executor callback interface . 
 
@@ -361,7 +361,7 @@ healthcheck: health check is enabled or not in your plugins. (Required)
 foldername: folder to keep temporary files generated by plugins. (Optional, default folder name is poddata)
 -->
 
-##### General Plugin configuration file(plugin/general/general.yaml)
+##### General Plugin configuration file(pluginimpl/general/general.yaml)
 Individual Plugin (such as General Plugin) configuration file caters to plugin relevant information. See details below:
 
 ```
