+++ pluginimpl/example/impl.go
