+++ pluginimpl/general/editor.go
