+++ pluginimpl/general/editor_test.go
