+++ pluginimpl/general/impl.go
