+++ pluginimpl/general/testdata/docker-infra-container.yml
